move TableMetaDataInitializerEntry to shardingsphere-common
