+++ shardingsphere-underlying/shardingsphere-common/src/main/java/org/apache/shardingsphere/underlying/common/metadata/table/init/TableMetaDataInitializerEntry.java
@@ -15,12 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.metadata;
+package org.apache.shardingsphere.underlying.common.metadata.table.init;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.underlying.common.metadata.table.TableMetaData;
 import org.apache.shardingsphere.underlying.common.metadata.table.TableMetas;
-import org.apache.shardingsphere.underlying.common.metadata.table.init.TableMetaDataInitializer;
 import org.apache.shardingsphere.underlying.common.metadata.table.init.decorator.TableMetaDataDecorator;
 import org.apache.shardingsphere.underlying.common.metadata.table.init.loader.TableMetaDataLoader;
 import org.apache.shardingsphere.underlying.common.rule.BaseRule;
